Clean: Merge STEX RasterOpGeneric
